[hotfix] #96 - ProductRegisterViewController 파일 분리
